refactor(tui): extract use-file-search hook (Issue 23, PR 4)
Move @-token detection useEffect, acceptPickerSelection handler, and detectAtToken from app.tsx to hooks/use-file-search.ts

app.tsx: -101 lines (13 added), hook: +144 lines
tests: import path fixed, PR 0 baseline expanded
